backend/pyproject.toml
[tool.poetry]
name = "google-ads-optimizer"
version = "0.1.0"
description = "Production-grade Google Ads copy analyzer and optimizer"
authors = ["Your Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
google-ads = "^26.0.0"
google-auth = "^2.26.2"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"
openai = "^1.10.0"
python-dotenv = "^1.0.1"
pydantic = {extras = ["email"], version = "^2.5.3"}
pydantic-settings = "^2.1.0"
cryptography = "^42.0.0"
httpx = "^0.26.0"
python-multipart = "^0.0.6"
itsdangerous = "^2.1.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
numpy = "^1.26.3"
scikit-learn = "^1.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.1"
ruff = "^0.1.14"
mypy = "^1.8.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
